refactor(PostList): Refactor
