+++ src/components/Post/index.tsx
@@ -1,3 +1,4 @@
 export * from "./Post";
 export * from "./PostDetail";
+export * from "./PostList";
 export * from "./PostPlaceholder";
