+++ src/pages/virtualized-list/[[...slug]].tsx
@@ -1,21 +1,15 @@
 import { motion } from "framer-motion";
 import type { GetServerSideProps } from "next";
 import { useRouter } from "next/router";
-import React, { useCallback, useMemo } from "react";
-import { AutoSizer, InfiniteLoader, List, ListRowRenderer, WindowScroller } from "react-virtualized";
-import 'react-virtualized/styles.css';
+import React, { useMemo } from "react";
 import { IS_SERVER } from "../../common/isServer";
 import { appBaseStyle, transition } from "../../components/layout";
 import { PageHeader } from "../../components/PageHeader";
-import { Post, PostDetail, PostPlaceholder } from "../../components/Post";
+import { PostDetail } from "../../components/Post";
 import { addApolloStateToPageProps, initializeApollo } from "../../hooks/useAplloClient";
 import { useVirtualizedList } from "../../hooks/VirtualizedList.hooks";
 import { GetPostConnectionDocument, GetPostConnectionQuery } from "../../__generated__/graphqlOperationTypes";
-
-const INFINITE_LOAD_THRESHOLD = 3;
-const INFINITE_LOAD_MIN_BATCH_SIZE = 1;
-const ROW_HEIGHT = 96;
-const ROW_MARGIN = 8;
+import { PostList } from "../../components/Post/PostList";
 
 type Props = {
   posts: GetPostConnectionQuery;
@@ -53,21 +47,6 @@ const VirtualizedList: React.FC<Props> = () => {
     return null;
   }, [postId, posts]);
 
-  const isRowLoaded = useCallback(({ index }: { index: number }) => {
-    return Boolean(posts[index]);
-  }, [posts]);
-
-  const rowRenderer: ListRowRenderer = useCallback(({ key, index, style }) => {
-    const post = posts[index];
-    return (
-      <div key={key} style={style}>
-        <div className="hover:bg-gray-50 cursor-pointer">
-          {post ? <Post post={post} /> : <PostPlaceholder />}
-        </div>
-      </div>
-    );
-  }, [posts]);
-
   return (
     <>
       <motion.div
@@ -79,43 +58,11 @@ const VirtualizedList: React.FC<Props> = () => {
       >
         <PageHeader title={"Virtualized List example"} />
         <div className="mt-9 mb-24">
-          <InfiniteLoader
-            isRowLoaded={isRowLoaded}
-            loadMoreRows={fetchMorePosts}
-            rowCount={totalCount}
-            threshold={INFINITE_LOAD_THRESHOLD}
-            minimumBatchSize={INFINITE_LOAD_MIN_BATCH_SIZE}
-          >
-            {({ onRowsRendered, registerChild }) => {
-              return (
-                <WindowScroller>
-                  {({ height, isScrolling, scrollTop, onChildScroll }) => {
-                    return (
-                      <AutoSizer disableHeight={true} >
-                        {({ width }) => {
-                          return (
-                            <List
-                              autoHeight
-                              height={height}
-                              width={width}
-                              isScrolling={isScrolling}
-                              scrollTop={scrollTop}
-                              onScroll={onChildScroll}
-                              onRowsRendered={onRowsRendered}
-                              ref={registerChild}
-                              rowCount={totalCount}
-                              rowHeight={ROW_HEIGHT + ROW_MARGIN}
-                              rowRenderer={rowRenderer}
-                            />
-                          );
-                        }}
-                      </AutoSizer>
-                    );
-                  }}
-                </WindowScroller>
-              );
-            }}
-          </InfiniteLoader>
+          <PostList
+            totalCount={totalCount}
+            posts={posts}
+            fetchMorePosts={fetchMorePosts}
+          />
         </div>
       </motion.div>
       {!IS_SERVER && postId && post && <PostDetail post={post} />}
